Fix @lexical/utils import (#5245)
